+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 All notable changes to `dash` will be documented in this file.
 This project adheres to [Semantic Versioning](https://semver.org/).
 
-## [UNRELEASED]
+## [2.13.0] 2023-08-28
 ## Changed
 
 - [#2610](https://github.com/plotly/dash/pull/2610) Load plotly.js bundle/version from plotly.py